Added expirations fetching from Tradier.
Changed Open Position form to show a drop-down list for expirations (fetched from Tradier). For now these are hard-coded to use the expirations for AAPL, as we don't know what symbol is being entered when the page is generated.

Converted expiration date from OffsetDateTime to LocalDate as we don't need the time information.